+++ src/main/java/net/frozenblock/lib/config/api/instance/xjs/XjsFormat.java
@@ -1,49 +1,34 @@
 /*
- * Copyright 2023 The Quilt Project
- * Copyright 2023 FrozenBlock
- * Modified to work on Fabric
+ * Copyright (C) 2024 FrozenBlock
  *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * ;;match_from: \/\*\r?\n \* Copyright (\(c\) )?2022 The Quilt Project
- * ;;match_from: \/\*\r?\n \* Copyright (\(c\) )?2021-2022 The Quilt Project
- * ;;match_from: \/\*\r?\n \* Copyright (\(c\) )?2023 The Quilt Project
- * ;;match_from: \/\*\r?\n \* Copyright (\(c\) )?2021-2023 The Quilt Project
- * ;;match_from: \/\*\r?\n \* Copyright (\(c\) )?2022-2023 The Quilt Project
- * ;;match_from: \/\*\r?\n \* Copyright (\(c\) )?2022 QuiltMC
- * ;;match_from: \/\*\r?\n \* Copyright (\(c\) )?2021-2022 QuiltMC
- * ;;match_from: \/\*\r?\n \* Copyright (\(c\) )?2023 QuiltMC
- * ;;match_from: \/\*\r?\n \* Copyright (\(c\) )?2021-2023 QuiltMC
- * ;;match_from: \/\*\r?\n \* Copyright (\(c\) )?2022-2023 QuiltMC
- * ;;match_from: \/\/\/ Q[Uu][Ii][Ll][Tt]
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
 package net.frozenblock.lib.config.api.instance.xjs;
 
-import net.minecraft.util.StringRepresentable;
-import org.jetbrains.annotations.NotNull;
-import xjs.serialization.util.UBTyping;
-import xjs.serialization.writer.HjsonWriter;
-import xjs.serialization.writer.JsonWriter;
-import xjs.serialization.writer.TxtWriter;
-import xjs.serialization.writer.UbjsonWriter;
-import xjs.serialization.writer.ValueWriter;
-import xjs.serialization.writer.XjsWriter;
 import java.io.File;
 import java.io.IOException;
-import java.io.Writer;
-import java.nio.file.Path;
 import java.util.function.Function;
+import net.minecraft.util.StringRepresentable;
+import org.jetbrains.annotations.NotNull;
+import xjs.compat.serialization.util.UBTyping;
+import xjs.compat.serialization.writer.HjsonWriter;
+import xjs.compat.serialization.writer.TxtWriter;
+import xjs.compat.serialization.writer.UbjsonWriter;
+import xjs.data.serialization.writer.DjsWriter;
+import xjs.data.serialization.writer.JsonWriter;
+import xjs.data.serialization.writer.ValueWriter;
 
 public enum XjsFormat implements StringRepresentable {
 	/**
@@ -91,22 +76,22 @@ public enum XjsFormat implements StringRepresentable {
 	}),
 
 	/**
-	 * Prints unformatted XJS with no whitespace.
+	 * Prints unformatted DJS with no whitespace.
 	 */
-	XJS("xjs", writer -> {
+	DJS("djs", writer -> {
 		try {
-			return new XjsWriter(writer, false);
+			return new DjsWriter(writer, false);
 		} catch (IOException e) {
 			throw new RuntimeException(e);
 		}
 	}),
 
 	/**
-	 * Pretty prints XJS with whitespace.
+	 * Pretty prints DJS with whitespace.
 	 */
-	XJS_FORMATTED(XJS.getSerializedName(), writer -> {
+	DJS_FORMATTED(DJS.getSerializedName(), writer -> {
 		try {
-			return new XjsWriter(writer, true);
+			return new DjsWriter(writer, true);
 		} catch (IOException e) {
 			throw new RuntimeException(e);
 		}